Added example of Long output to README

+++ main.py
@@ -9,7 +9,7 @@ def main():
     search_location = 'roppongi hills mori tower'
     search_item = "cafe"
     radius = 200
-    output_result = 'Short'
+    output_result = 'Long'
 
     client = GoogleMapsClient(api_key=API_key,address= search_location)
     search_nearby_results = client.search_nearby(search_item,radius=radius,output_result=output_result)
